tests: Add stress test exe
