+++ tests/CMakeLists.txt
@@ -326,12 +326,10 @@ if(DEFINED SLANG_INSTALL_DIR)
         # Find libslang.so
         if(EXISTS "${SLANG_LIB_DIR}/libslang.so")
             set(SLANG_LIBRARY "${SLANG_LIB_DIR}/libslang.so")
-            message(STATUS "Found libslang.so at: ${SLANG_LIBRARY}")
         else()
             file(GLOB SLANG_LIB_FILES "${SLANG_INSTALL_DIR}/**/libslang.so")
             if(SLANG_LIB_FILES)
                 list(GET SLANG_LIB_FILES 0 SLANG_LIBRARY)
-                message(STATUS "Found libslang.so via glob at: ${SLANG_LIBRARY}")
             else()
                 message(FATAL_ERROR "Could not locate libslang.so in extracted directory")
             endif()
@@ -480,5 +478,6 @@ if(NOT VVL_SKIP_GTEST_DISCOVERY)
     gtest_discover_tests(vk_layer_validation_tests DISCOVERY_TIMEOUT 100)
 endif()
 
+add_subdirectory(stress)
 add_subdirectory(layers)
 add_subdirectory(icd)
